Rename module to better reflect purpose
