+++ wsgi.py
@@ -1,4 +1,4 @@
-from src.main import create_app
+from src.app_factory import create_app
 
 app = create_app()
